Fix ruff RUF012/RUF046 findings
A redundant int() cast and two mutable class-attribute defaults.

+++ custom_components/gardena_smart_local_preview/number.py
@@ -93,7 +93,7 @@ class GardenaButtonConfigTime(GardenaEntity, NumberEntity):
         seconds = device.button_config_time
         if seconds is None:
             return None
-        return int(round(seconds / 60))
+        return round(seconds / 60)
 
     async def async_set_native_value(self, value: float) -> None:
         await self.coordinator.send_request(
